refactor: reorganize project structure into modules and implement event-driven app architecture
